+++ src/components/Cocina.js
@@ -1,0 +1,20 @@
+/* eslint-disable react/prop-types */
+import React, { Fragment } from 'react'
+import { Link } from 'react-router-dom';
+
+export const Cocina = () => {
+   
+    return (
+       <Fragment>
+           <div className="cocina">
+            <h1>Pedido en cocina</h1> 
+        </div>
+
+        <Link to='/Login'> :</Link>
+
+       </Fragment>
+        
+
+        
+    );
+}
