+++ src/components/Container.js
@@ -5,21 +5,19 @@ import logo from '../assets/logo.png';
 import table from '../assets/table.png';
 import Menu from '../components/Menu';
 import helpHttp from '../helpers/Helpers';
+import Comanda from '../components/Comanda';
+
 
 
 export const Container = function () {
 	const [error, setError] = useState('');
-	const [nombre, setNombre] = useState('')
 	const { logout } = useAuth();
 	const [menu, setMenu] = useState('desayuno');
 	const history = useHistory();
+	
 	const [data, setData] = useState([]);
-
-	// const [dataToEdit, setDataToEdit] = useState(null);
 	let api = helpHttp();
 	let url = ' http://localhost:5000/Products';
-
-	// let [products, setProducts] = useState();
 
 	useEffect(() => {
 		api.get(url).then((res) => {
@@ -40,14 +38,13 @@ export const Container = function () {
 		}
 	}
 
-	const handleName = e => setNombre(e.target.value);
+	// const handleName = e => setNombre(e.target.value);
 
 	const filterProductos = () => {
 		return data.filter((p) => p.type == menu);
 	};
 
 	
-
 	return (
 		<Fragment>
 			<header >
@@ -59,15 +56,12 @@ export const Container = function () {
 						<span className="material-icons">
 							face
 						</span>
+						
 						<label>
 							Nombre del cliente  :
-							<input type="nombre" className="input-name-cliente" placeholder='Ingresa nombre del cliente' onChange={handleName} required />
+							<input type="nombre" className="input-name-cliente" placeholder='' required />
 						</label>
-						<input type='submit' data-testid='submit-test' className='btn-nombre' value='Ingresar el nombre a comanda'
-							onClick={(e) => {
-								e.preventDefault();
-								handleName(nombre);
-							}} />
+						<button className="btn-nom-cliente"> Ingresa Nombre de Cliente a Comanda</button>
 					</div>
 					<div>
 						<span className="material-icons" onClick={handleLogout}>
@@ -85,30 +79,22 @@ export const Container = function () {
 					<div className="btns-menu">
 						<button onClick={() => { setMenu("desayuno") }} className="btn-desayuno">DESAYUNO</button>
 						<button onClick={() => { setMenu("almuerzo") }} className="btn-almuerzo">ALMUERZO</button>
-						<div>
-							{data && <Menu products={filterProductos()} />}
-						</div>
 					</div>
-					{/* {products && products.map(product =>
-						<Menu product={product} key={product.id} />
-					)} */}
+					<div className="table-products">
+							{data && <Menu products={filterProductos()} />}
+					</div>
 				</div>
 
+				{/* COMPONENTE COMANDA */}
+				<Comanda/>
 
-				{/* CODIGO COMANDA */}
-				<div className="comanda">
-					<h2>Comanda</h2>
-					<button className="btn-enviar-cocina">ENVIAR A COCINA</button>
-					<p>Cliente estrella : {handleName}</p>
-
-				</div>
-			</div>
+			</div> 
 
 			<footer>
 				<br />
 				<br />
 				<br />
-				<h4>Hola ( Nombre Usuario ), estas en el sistema de pedidos</h4>
+				<h4>Hola, estas en el sistema de pedidos</h4>
 				<img src={table} alt='table' className="footer-table" />
 			</footer>
 		</Fragment>
