--- src/components/CrudApi.js
@@ -1,9 +1,0 @@
-const CrudApi = () => {
-    return (
-        <div>
-            <h2>Crud API</h2>
-        </div>
-    )
-}
-
-export default CrudApi;
